Move sample codes out of SurveyController & Hide sample codes from swagger
